delete restriction

+++ flaskr/blog.py
@@ -474,9 +474,12 @@ def tag(page=None,tag_slug=None):
 @login_required
 def delete(id):
     db = get_db()
-    db.execute('DELETE FROM post_tag where post_id = %s;', (id,))
-    db.execute('DELETE FROM post_comment where post_id = %s;', (id,))
-    db.execute('DELETE FROM post WHERE id = %s;', (id,))
+    post_user = db.execute('SELECT id FROM post WHERE id = %s;', (id,)).fetchone()[0]
+    # only delete if same user, or if admin 
+    if g.user == post_user or g.user['role_id'] == 2:
+        db.execute('DELETE FROM post_tag where post_id = %s;', (id,))
+        db.execute('DELETE FROM post_comment where post_id = %s;', (id,))
+        db.execute('DELETE FROM post WHERE id = %s;', (id,))
     return redirect(url_for('blog.index'))
     
 @bp.route('/<int:id>/comment_delete', methods=('POST',))
